Add files via upload
First screenshot is top level block diagram, followed by zoomed in pictures for better legibility for reading separate hardware blocks. The code for VGA controller and image generator are on separate files contained in the repository. This diagram was mainly borrowed from our resource which is linked under resources on the wiki.